requirements.txt
openai~=1.12.0
PyYAML~=6.0.1
requests~=2.31.0
langchain~=0.1.7
langchain_openai~=0.0.6
pydantic~=2.6.1
numpy~=1.26.4